+++ lua/cmp_codecompanion/slash_commands.lua
@@ -1,5 +1,5 @@
-local config = require("codecompanion").config
 local SlashCommands = require("codecompanion.strategies.chat.slash_commands")
+local config = require("codecompanion.config")
 local strategy = require("codecompanion.strategies")
 
 local source = {}
@@ -81,9 +81,9 @@ function source:execute(item, callback)
   if item.from_prompt_library then
     local prompts = strategy.evaluate_prompts(item.config.prompts, item.context)
     vim.iter(prompts):each(function(prompt)
-      if prompt.role == "system" then
+      if prompt.role == config.constants.SYSTEM_ROLE then
         item.Chat:add_message(prompt, { visible = false })
-      elseif prompt.role == "user" then
+      elseif prompt.role == config.constants.USER_ROLE then
         item.Chat:append_to_buf(prompt)
       end
     end)