+++ lua/codecompanion/config.lua
@@ -1,6 +1,12 @@
 local fmt = string.format
 
-return {
+local constants = {
+  LLM_ROLE = "llm",
+  USER_ROLE = "user",
+  SYSTEM_ROLE = "system",
+}
+
+local defaults = {
   adapters = {
     -- LLMs -------------------------------------------------------------------
     anthropic = "anthropic",
@@ -312,7 +318,7 @@ Points to note:
       },
       prompts = {
         {
-          role = "system",
+          role = constants.SYSTEM_ROLE,
           content = function(context)
             return fmt(
               [[I want you to act as a senior %s developer. I will ask you specific questions and I want you to return raw code only (no codeblocks and no explanations). If you can't respond with code, respond with nothing]],
@@ -341,7 +347,7 @@ Points to note:
       },
       prompts = {
         {
-          role = "system",
+          role = constants.SYSTEM_ROLE,
           content = [[When asked to explain code, follow these steps:
 
 1. Identify the programming language.
@@ -354,7 +360,7 @@ Points to note:
           },
         },
         {
-          role = "user",
+          role = constants.USER_ROLE,
           content = function(context)
             local code = require("codecompanion.helpers.actions").get_code(context.start_line, context.end_line)
 
@@ -391,7 +397,7 @@ Points to note:
       },
       prompts = {
         {
-          role = "system",
+          role = constants.SYSTEM_ROLE,
           content = [[When generating unit tests, follow these steps:
 
 1. Identify the programming language.
@@ -408,7 +414,7 @@ Points to note:
           },
         },
         {
-          role = "user",
+          role = constants.USER_ROLE,
           content = function(context)
             local code = require("codecompanion.helpers.actions").get_code(context.start_line, context.end_line)
 
@@ -445,7 +451,7 @@ Points to note:
       },
       prompts = {
         {
-          role = "system",
+          role = constants.SYSTEM_ROLE,
           content = [[When asked to fix code, follow these steps:
 
 1. **Identify the Issues**: Carefully read the provided code and identify any potential issues or improvements.
@@ -466,7 +472,7 @@ Use Markdown formatting and include the programming language name at the start o
           },
         },
         {
-          role = "user",
+          role = constants.USER_ROLE,
           content = function(context)
             local code = require("codecompanion.helpers.actions").get_code(context.start_line, context.end_line)
 
@@ -503,7 +509,7 @@ Use Markdown formatting and include the programming language name at the start o
       },
       prompts = {
         {
-          role = "system",
+          role = constants.SYSTEM_ROLE,
           content = function(context)
             return "I want you to act as a senior "
               .. context.filetype
@@ -515,7 +521,7 @@ Use Markdown formatting and include the programming language name at the start o
           },
         },
         {
-          role = "user",
+          role = constants.USER_ROLE,
           content = function(context)
             local buf_utils = require("codecompanion.utils.buffers")
 
@@ -527,7 +533,7 @@ Use Markdown formatting and include the programming language name at the start o
           },
         },
         {
-          role = "user",
+          role = constants.USER_ROLE,
           condition = function(context)
             return context.is_visual
           end,
@@ -568,14 +574,14 @@ Use Markdown formatting and include the programming language name at the start o
       },
       prompts = {
         {
-          role = "system",
+          role = constants.SYSTEM_ROLE,
           content = [[You are an expert coder and helpful assistant who can help debug code diagnostics, such as warning and error messages. When appropriate, give solutions with code snippets as fenced codeblocks with a language identifier to enable syntax highlighting.]],
           opts = {
             visible = false,
           },
         },
         {
-          role = "user",
+          role = constants.USER_ROLE,
           content = function(context)
             local diagnostics = require("codecompanion.helpers.actions").get_diagnostics(
               context.start_line,
@@ -611,7 +617,7 @@ Use Markdown formatting and include the programming language name at the start o
           end,
         },
         {
-          role = "user",
+          role = constants.USER_ROLE,
           content = function(context)
             local code = require("codecompanion.helpers.actions").get_code(
               context.start_line,
@@ -648,7 +654,7 @@ This is the code, for context:
       },
       prompts = {
         {
-          role = "user",
+          role = constants.USER_ROLE,
           content = function()
             return fmt(
               [[You are an expert at following the Conventional Commit specification. Given the git diff listed below, please generate a commit message for me:
@@ -768,3 +774,22 @@ When given a task:
 4. You can only give one reply for each conversation turn.]],
   },
 }
+
+local M = {
+  config = vim.deepcopy(defaults),
+}
+
+---@param args? table
+M.setup = function(args)
+  args = args or {}
+  M.config = vim.tbl_deep_extend("force", vim.deepcopy(defaults), { constants = vim.deepcopy(constants) }, args)
+end
+
+return setmetatable(M, {
+  __index = function(_, key)
+    if key == "setup" then
+      return M.setup
+    end
+    return rawget(M.config, key)
+  end,
+})