+++ lua/codecompanion/strategies/chat/tools.lua
@@ -1,5 +1,5 @@
 local Job = require("plenary.job")
-local config = require("codecompanion").config
+local config = require("codecompanion.config")
 
 local TreeHandler = require("codecompanion.utils.xml.xmlhandler.tree")
 local log = require("codecompanion.utils.log")
@@ -17,10 +17,6 @@ local CONSTANTS = {
 
   STATUS_ERROR = "error",
   STATUS_SUCCESS = "success",
-
-  USER_ROLE = "user",
-  LLM_ROLE = "llm",
-  SYSTEM_ROLE = "system",
 
   PROCESSING_MSG = "Tool processing ...",
 }
@@ -103,7 +99,7 @@ function Tools:set_autocmds()
         if request.data.status == CONSTANTS.STATUS_ERROR and self.tool.output_error_prompt then
           log:debug("Tool %s finished with error", self.tool.name)
           self.chat:append_to_buf({
-            role = CONSTANTS.USER_ROLE,
+            role = config.constants.USER_ROLE,
             content = self.tool.output_error_prompt(request.data.stderr),
           })
           if self.tools_config.opts.auto_submit_errors then
@@ -118,14 +114,14 @@ function Tools:set_autocmds()
           log:debug("Tool output: %s", output)
 
           local message = {
-            role = CONSTANTS.USER_ROLE,
+            role = config.constants.USER_ROLE,
             content = self.tool.output_prompt(output),
           }
 
           if self.tools_config[self.tool.name].opts and self.tools_config[self.tool.name].opts.hide_output then
             self.chat:add_message(message, { visible = false })
             self.chat:append_to_buf({
-              role = CONSTANTS.USER_ROLE,
+              role = config.constants.USER_ROLE,
               content = "I've shared the output with you",
             })
           else
@@ -273,6 +269,10 @@ end
 ---@param message table
 ---@return table|nil
 function Tools:find(message)
+  if not message.content then
+    return nil
+  end
+
   local found = {}
   for tool, _ in pairs(self.tools_config) do
     if message.content:match("%f[%w" .. CONSTANTS.PREFIX .. "]" .. CONSTANTS.PREFIX .. tool .. "%f[%W]") then
@@ -359,13 +359,13 @@ end
 ---@return CodeCompanion.Tools
 function Tools:add_error_to_chat(error)
   self.chat:add_message({
-    role = CONSTANTS.USER_ROLE,
+    role = config.constants.USER_ROLE,
     content = error,
   }, { visible = false })
 
   --- Alert the user that the error message has been shared
   self.chat:append_to_buf({
-    role = CONSTANTS.USER_ROLE,
+    role = config.constants.USER_ROLE,
     content = "Please correct for the error message I've shared",
   })